+++ packages/sdks/e2e/tests/src/tests/main.spec.ts
@@ -1,4 +1,4 @@
-import type { Locator, Page } from '@playwright/test';
+import type { BrowserContext, Locator, Page } from '@playwright/test';
 import { test, expect } from '@playwright/test';
 
 import { targetContext } from './context.js';
@@ -8,6 +8,8 @@ import { targetContext } from './context.js';
 const findTextInPage = ({ page, text }: { page: Page; text: string }) =>
   expect(page.locator(`text=${text}`)).toBeVisible();
 
+const isRNSDK = process.env.SDK === 'reactNative';
+
 const getElementStyleValue = async ({
   locator,
   cssProperty,
@@ -15,18 +17,87 @@ const getElementStyleValue = async ({
   locator: Locator;
   cssProperty: string;
 }) => {
-  return locator.evaluate(
-    (e, cssProperty) => getComputedStyle(e).getPropertyValue(cssProperty),
-    cssProperty
+  return locator.evaluate((e, cssProperty) => {
+    return getComputedStyle(e).getPropertyValue(cssProperty);
+  }, cssProperty);
+};
+
+const expectStyleForElement = async ({
+  expectedValue,
+  locator,
+  cssProperty,
+}: {
+  locator: Locator;
+  cssProperty: string;
+  expectedValue: string;
+}) => {
+  await expect(
+    await getElementStyleValue({
+      locator,
+      cssProperty,
+    })
+  ).toBe(expectedValue);
+};
+const expectStylesForElement = async ({
+  expected,
+  locator,
+}: {
+  locator: Locator;
+  expected: Record<string, string>;
+}) => {
+  for (const property of Object.keys(expected)) {
+    await expectStyleForElement({
+      cssProperty: property,
+      locator,
+      expectedValue: expected[property],
+    });
+  }
+};
+
+const getBuilderSessionIdCookie = async ({
+  context,
+}: {
+  context: BrowserContext;
+}) => {
+  const cookies = await context.cookies();
+  const builderSessionCookie = cookies.find(
+    (cookie) => cookie.name === 'builderSessionId'
   );
+  return builderSessionCookie;
 };
 
 test.describe(targetContext.name, () => {
+  test.describe('cookies', () => {
+    test('do not appear if canTrack=false', async ({ page, context }) => {
+      // by waiting for network requests, we guarantee that impression tracking POST was (NOT) made,
+      // which guarantees that the cookie was set or not.
+      await page.goto('/can-track-false', { waitUntil: 'networkidle' });
+
+      const cookies = await context.cookies();
+      const builderSessionCookie = cookies.find(
+        (cookie) => cookie.name === 'builderSessionId'
+      );
+      expect(builderSessionCookie).toBeUndefined();
+    });
+    test('appear by default', async ({ page, context }) => {
+      // by waiting for network requests, we guarantee that impression tracking POST was made,
+      // which guarantees that the cookie was set
+      await page.goto('/', { waitUntil: 'networkidle' });
+
+      const builderSessionCookie = await getBuilderSessionIdCookie({ context });
+
+      // react native sdk does not use cookies
+      if (isRNSDK) {
+        expect(builderSessionCookie).toBeUndefined();
+      } else {
+        expect(builderSessionCookie).toBeDefined();
+      }
+    });
+  });
   test('homepage', async ({ page }) => {
     await page.goto('/');
 
     const links = await page.locator('a');
-    await expect(links).toHaveCount(6);
 
     const columnsLink = await links.filter({
       hasText: 'Columns (with images) ',
@@ -56,6 +127,95 @@ test.describe(targetContext.name, () => {
         '[src="https://cdn.builder.io/api/v1/image/assets%2Ff1a790f8c3204b3b8c5c1795aeac4660%2F4bce19c3d8f040b3a95e91000a98283e"]'
       )
       .isVisible();
+
+    const firstSymbolText = await page
+      .locator('text="Description of image:"')
+      .first();
+
+    // these are desktop and tablet styles, and will never show up in react native
+    if (!isRNSDK) {
+      // check desktop styles
+      await expectStyleForElement({
+        locator: firstSymbolText,
+        cssProperty: 'color',
+        expectedValue: 'rgb(255, 0, 0)',
+      });
+
+      // resize to tablet
+      await page.setViewportSize({ width: 930, height: 1000 });
+      await expectStyleForElement({
+        locator: firstSymbolText,
+        cssProperty: 'color',
+        expectedValue: 'rgb(0, 255, 6)',
+      });
+
+      // resize to mobile
+      await page.setViewportSize({ width: 400, height: 1000 });
+    }
+
+    // TO-DO: fix react native style inheritance for symbols->Text (using HTML renderer component), so we can unblock this.
+    if (!isRNSDK) {
+      // check mobile styles
+      await expectStyleForElement({
+        locator: firstSymbolText,
+        cssProperty: 'color',
+        expectedValue: 'rgb(0, 255, 255)',
+      });
+    }
+  });
+  test('style bindings', async ({ page }) => {
+    await page.goto('/content-bindings');
+
+    const expected = {
+      'border-top-left-radius': '10px',
+      'border-top-right-radius': '22px',
+      'border-bottom-left-radius': '40px',
+      'border-bottom-right-radius': '30px',
+    };
+
+    const selector = isRNSDK
+      ? '[data-class*=builder-blocks] > div'
+      : '[class*=builder-blocks] > div';
+
+    const locator = page
+      .locator(selector)
+      .filter({ hasText: 'Enter some text...' })
+      .last();
+
+    page.locator(selector).innerText;
+
+    await expect(locator).toBeVisible();
+
+    await expectStylesForElement({ expected, locator });
+    // TODO: fix this
+    // check the title is correct
+    // title: 'some special title'
+  });
+  test('symbol style bindings', async ({ page }) => {
+    await page.goto('/symbol-bindings');
+
+    const expected = {
+      'border-top-left-radius': '10px',
+      'border-top-right-radius': '220px',
+      'border-bottom-left-radius': '30px',
+      'border-bottom-right-radius': '40px',
+    };
+
+    const selector = isRNSDK
+      ? '[data-class*=builder-blocks] > div'
+      : '[class*=builder-blocks] > div';
+
+    const locator = page
+      .locator(selector)
+      .filter({ hasText: 'Enter some text...' })
+      .last();
+
+    await expect(locator).toBeVisible();
+
+    await expectStylesForElement({ expected, locator });
+    // TODO: fix this
+    // check the title is correct
+    // title: 'some special title'
   });
   test('data-bindings', async ({ page }) => {
     await page.goto('/data-bindings');
@@ -70,15 +230,29 @@ test.describe(targetContext.name, () => {
       text: 'Mattel Certified by Great Place to Work and Named to Fast Company’s List of 100 Best Workplaces for Innovators',
     });
   });
+  test('data-binding-styles', async ({ page }) => {
+    await page.goto('/data-binding-styles');
+    if (isRNSDK) {
+      // styling is not yet implemented in RN SDK
+      return;
+    }
+    await expectStyleForElement({
+      locator: page.locator(`text="This text should be red..."`),
+      cssProperty: 'color',
+      expectedValue: 'rgb(255, 0, 0)',
+    });
+  });
 
-  test('image', async ({ page }) => {
+  /**
+   * We are temporarily skipping this test because it relies on network requests.
+   * TO-DO: re-enable it once we have a way to mock network requests.
+   */
+  test.skip('image', async ({ page }) => {
     await page.goto('/image');
 
     const imageLocator = await page.locator('img');
 
-    const isRNSDK = process.env.SDK === 'reactNative';
-
-    const expected = [
+    const expected: Record<string, string>[] = [
       // first img is a webp image. React Native SDK does not yet support webp.
       ...(isRNSDK
         ? []
@@ -111,14 +285,7 @@ test.describe(targetContext.name, () => {
 
     expected.forEach(async (vals, index) => {
       const image = imageLocator.nth(index);
-      for (const [key, value] of Object.entries(vals)) {
-        await expect(
-          await getElementStyleValue({
-            locator: image,
-            cssProperty: key,
-          })
-        ).toBe(value);
-      }
+      await expectStylesForElement({ locator: image, expected: vals });
     });
   });
 
@@ -142,12 +309,11 @@ test.describe(targetContext.name, () => {
           expectedTextColor = 'rgb(65, 117, 5)'; // greenish text color
         }
 
-        await expect(
-          await getElementStyleValue({
-            locator: breakpointsPara,
-            cssProperty: 'color',
-          })
-        ).toBe(expectedTextColor);
+        await expectStyleForElement({
+          locator: breakpointsPara,
+          cssProperty: 'color',
+          expectedValue: expectedTextColor,
+        });
 
         const column2 = page.locator(`text=Column 2`);
         await expect(column2).toBeVisible();
@@ -157,12 +323,11 @@ test.describe(targetContext.name, () => {
           expectedColumnTextColor = 'rgb(126, 211, 33)'; // greenish text color
         }
 
-        await expect(
-          await getElementStyleValue({
-            locator: column2,
-            cssProperty: 'color',
-          })
-        ).toBe(expectedColumnTextColor);
+        await expectStyleForElement({
+          locator: column2,
+          cssProperty: 'color',
+          expectedValue: expectedColumnTextColor,
+        });
 
         // Skipping this image test for react-native.
         // Its difficult to locate the image in react-native as css selectors don't work as expected.
@@ -175,14 +340,10 @@ test.describe(targetContext.name, () => {
             width: '785px',
           };
 
-          for (const property of Object.keys(expectedImageCss)) {
-            await expect(
-              await getElementStyleValue({
-                locator: image,
-                cssProperty: property,
-              })
-            ).toBe(expectedImageCss[property]);
-          }
+          await expectStylesForElement({
+            locator: image,
+            expected: expectedImageCss,
+          });
         }
       });
 
@@ -198,12 +359,11 @@ test.describe(targetContext.name, () => {
           expectedTextColor = 'rgb(65, 117, 5)'; // greenish text color
         }
 
-        await expect(
-          await getElementStyleValue({
-            locator: breakpointsPara,
-            cssProperty: 'color',
-          })
-        ).toBe(expectedTextColor);
+        await expectStyleForElement({
+          locator: breakpointsPara,
+          cssProperty: 'color',
+          expectedValue: expectedTextColor,
+        });
 
         const column2 = page.locator(`text=Column 2`);
         await expect(column2).toBeVisible();
@@ -213,12 +373,11 @@ test.describe(targetContext.name, () => {
           expectedColumnTextColor = 'rgb(126, 211, 33)'; // greenish text color
         }
 
-        await expect(
-          await getElementStyleValue({
-            locator: column2,
-            cssProperty: 'color',
-          })
-        ).toBe(expectedColumnTextColor);
+        await expectStyleForElement({
+          locator: column2,
+          cssProperty: 'color',
+          expectedValue: expectedColumnTextColor,
+        });
 
         // Skipping this image test for react-native.
         // Its difficult to locate the image in react-native as css selectors don't work as expected.
@@ -230,14 +389,10 @@ test.describe(targetContext.name, () => {
             display: 'none',
           };
 
-          for (const property of Object.keys(expectedImageCss)) {
-            await expect(
-              await getElementStyleValue({
-                locator: image,
-                cssProperty: property,
-              })
-            ).toBe(expectedImageCss[property]);
-          }
+          await expectStylesForElement({
+            locator: image,
+            expected: expectedImageCss,
+          });
         }
       });
 
@@ -247,22 +402,20 @@ test.describe(targetContext.name, () => {
 
         const breakpointsPara = page.locator(`text=BREAKPOINTS 500 - 800`);
         await expect(breakpointsPara).toBeVisible();
-        await expect(
-          await getElementStyleValue({
-            locator: breakpointsPara,
-            cssProperty: 'color',
-          })
-        ).toBe('rgb(65, 117, 5)'); // greenish text color
+        await expectStyleForElement({
+          locator: breakpointsPara,
+          cssProperty: 'color',
+          expectedValue: 'rgb(65, 117, 5)',
+        });
 
         const column2 = page.locator(`text=Column 2`);
         await expect(column2).toBeVisible();
 
-        await expect(
-          await getElementStyleValue({
-            locator: column2,
-            cssProperty: 'color',
-          })
-        ).toBe('rgb(126, 211, 33)'); // greenish text color
+        await expectStyleForElement({
+          locator: column2,
+          cssProperty: 'color',
+          expectedValue: 'rgb(126, 211, 33)', // greenish text color
+        });
 
         // Skipping this image test for react-native.
         // Its difficult to locate the image in react-native as css selectors don't work as expected.
@@ -276,14 +429,10 @@ test.describe(targetContext.name, () => {
             'max-width': '250px',
           };
 
-          for (const property of Object.keys(expectedImageCss)) {
-            await expect(
-              await getElementStyleValue({
-                locator: image,
-                cssProperty: property,
-              })
-            ).toBe(expectedImageCss[property]);
-          }
+          await expectStylesForElement({
+            locator: image,
+            expected: expectedImageCss,
+          });
         }
       });
     });
@@ -308,12 +457,11 @@ test.describe(targetContext.name, () => {
           expectedTextColor = 'rgb(65, 117, 5)'; // greenish text color
         }
 
-        await expect(
-          await getElementStyleValue({
-            locator: breakpointsPara,
-            cssProperty: 'color',
-          })
-        ).toBe(expectedTextColor); // black text color
+        await expectStyleForElement({
+          locator: breakpointsPara,
+          cssProperty: 'color',
+          expectedValue: expectedTextColor, // black text color
+        });
 
         const column2 = page.locator(`text=Column 2`);
         await expect(column2).toBeVisible();
@@ -323,12 +471,11 @@ test.describe(targetContext.name, () => {
           expectedColumnTextColor = 'rgb(126, 211, 33)'; // greenish text color
         }
 
-        await expect(
-          await getElementStyleValue({
-            locator: column2,
-            cssProperty: 'color',
-          })
-        ).toBe(expectedColumnTextColor);
+        await expectStyleForElement({
+          locator: column2,
+          cssProperty: 'color',
+          expectedValue: expectedColumnTextColor,
+        });
 
         // Skipping this image test for react-native.
         // Its difficult to locate the image in react-native as css selectors don't work as expected.
@@ -341,14 +488,10 @@ test.describe(targetContext.name, () => {
             width: '976px',
           };
 
-          for (const property of Object.keys(expectedImageCss)) {
-            await expect(
-              await getElementStyleValue({
-                locator: image,
-                cssProperty: property,
-              })
-            ).toBe(expectedImageCss[property]);
-          }
+          await expectStylesForElement({
+            locator: image,
+            expected: expectedImageCss,
+          });
         }
       });
 
@@ -364,12 +507,11 @@ test.describe(targetContext.name, () => {
           expectedTextColor = 'rgb(65, 117, 5)'; // greenish text color
         }
 
-        await expect(
-          await getElementStyleValue({
-            locator: breakpointsPara,
-            cssProperty: 'color',
-          })
-        ).toBe(expectedTextColor);
+        await expectStyleForElement({
+          locator: breakpointsPara,
+          cssProperty: 'color',
+          expectedValue: expectedTextColor,
+        });
 
         const column2 = page.locator(`text=Column 2`);
         await expect(column2).toBeVisible();
@@ -379,12 +521,11 @@ test.describe(targetContext.name, () => {
           expectedColumnTextColor = 'rgb(126, 211, 33)'; // greenish text color
         }
 
-        await expect(
-          await getElementStyleValue({
-            locator: column2,
-            cssProperty: 'color',
-          })
-        ).toBe(expectedColumnTextColor);
+        await expectStyleForElement({
+          locator: column2,
+          cssProperty: 'color',
+          expectedValue: expectedColumnTextColor,
+        });
 
         // Skipping this image test for react-native.
         // Its difficult to locate the image in react-native as css selectors don't work as expected.
@@ -396,14 +537,10 @@ test.describe(targetContext.name, () => {
             display: 'none',
           };
 
-          for (const property of Object.keys(expectedImageCss)) {
-            await expect(
-              await getElementStyleValue({
-                locator: image,
-                cssProperty: property,
-              })
-            ).toBe(expectedImageCss[property]);
-          }
+          await expectStylesForElement({
+            locator: image,
+            expected: expectedImageCss,
+          });
         }
       });
 
@@ -414,22 +551,20 @@ test.describe(targetContext.name, () => {
         const breakpointsPara = page.locator(`text=BREAKPOINTS 500 - 800`);
         await expect(breakpointsPara).toBeVisible();
 
-        await expect(
-          await getElementStyleValue({
-            locator: breakpointsPara,
-            cssProperty: 'color',
-          })
-        ).toBe('rgb(65, 117, 5)'); // greenish text color
+        await expectStyleForElement({
+          locator: breakpointsPara,
+          cssProperty: 'color',
+          expectedValue: 'rgb(65, 117, 5)', // greenish text color
+        });
 
         const column2 = page.locator(`text=Column 2`);
         await expect(column2).toBeVisible();
 
-        await expect(
-          await getElementStyleValue({
-            locator: column2,
-            cssProperty: 'color',
-          })
-        ).toBe('rgb(126, 211, 33)'); // greenish text color
+        await expectStyleForElement({
+          locator: column2,
+          cssProperty: 'color',
+          expectedValue: 'rgb(126, 211, 33)', // greenish text color
+        });
 
         // Skipping this image test for react-native.
         // Its difficult to locate the image in react-native as css selectors don't work as expected.
@@ -443,14 +578,10 @@ test.describe(targetContext.name, () => {
             'max-width': '250px',
           };
 
-          for (const property of Object.keys(expectedImageCss)) {
-            await expect(
-              await getElementStyleValue({
-                locator: image,
-                cssProperty: property,
-              })
-            ).toBe(expectedImageCss[property]);
-          }
+          await expectStylesForElement({
+            locator: image,
+            expected: expectedImageCss,
+          });
         }
       });
     });