+++ packages/sdks/src/blocks/symbol/symbol.lite.tsx
@@ -5,6 +5,7 @@ import type { BuilderContent } from '../../types/builder-content.js';
 import { onUpdate, useContext, useStore } from '@builder.io/mitosis';
 import type { Nullable } from '../../types/typescript.js';
 import type { BuilderBlock } from '../../types/builder-block.js';
+import { TARGET } from '../../constants/target';
 
 export interface SymbolInfo {
   model?: string;
@@ -28,7 +29,20 @@ export default function Symbol(props: SymbolProps) {
   const builderContext = useContext(BuilderContext);
 
   const state = useStore({
-    className: 'builder-symbol',
+    get className() {
+      return [
+        ...(TARGET === 'vue2' || TARGET === 'vue3'
+          ? Object.keys(props.attributes.class)
+          : [props.attributes.class]),
+        'builder-symbol',
+        props.symbol?.inline ? 'builder-inline-symbol' : undefined,
+        props.symbol?.dynamic || props.dynamic
+          ? 'builder-dynamic-symbol'
+          : undefined,
+      ]
+        .filter(Boolean)
+        .join(' ');
+    },
     fetchedContent: null as Nullable<BuilderContent>,
     get contentToUse(): Nullable<BuilderContent> {
       return props.symbol?.content || state.fetchedContent;