+++ packages/sdks/src/functions/get-content/index.ts
@@ -1,64 +1,14 @@
-import { flatten } from '../../helpers/flatten.js';
 import type { BuilderContent } from '../../types/builder-content.js';
-import {
-  getBuilderSearchParamsFromWindow,
-  normalizeSearchParams,
-} from '../get-builder-search-params/index.js';
-import { getFetch } from '../get-fetch.js';
+import { fetch } from '../get-fetch.js';
 import { handleABTesting } from './ab-testing.js';
-
-export type GetContentOptions = import('./types.js').GetContentOptions;
+import { generateContentUrl } from './generate-content-url.js';
+import type { GetContentOptions } from './types.js';
 
 export async function getContent(
   options: GetContentOptions
 ): Promise<BuilderContent | null> {
   return (await getAllContent({ ...options, limit: 1 })).results[0] || null;
 }
-
-export const generateContentUrl = (options: GetContentOptions): URL => {
-  const {
-    limit = 30,
-    userAttributes,
-    query,
-    noTraverse = false,
-    model,
-    apiKey,
-    includeRefs = true,
-    locale,
-  } = options;
-
-  if (!apiKey) {
-    throw new Error('Missing API key');
-  }
-
-  const url = new URL(
-    `https://cdn.builder.io/api/v2/content/${model}?apiKey=${apiKey}&limit=${limit}&noTraverse=${noTraverse}&includeRefs=${includeRefs}${
-      locale ? `&locale=${locale}` : ''
-    }`
-  );
-
-  const queryOptions = {
-    ...getBuilderSearchParamsFromWindow(),
-    ...normalizeSearchParams(options.options || {}),
-  };
-
-  const flattened = flatten(queryOptions);
-  for (const key in flattened) {
-    url.searchParams.set(key, String(flattened[key]));
-  }
-
-  if (userAttributes) {
-    url.searchParams.set('userAttributes', JSON.stringify(userAttributes));
-  }
-  if (query) {
-    const flattened = flatten({ query });
-    for (const key in flattened) {
-      url.searchParams.set(key, JSON.stringify((flattened as any)[key]));
-    }
-  }
-
-  return url;
-};
 
 /**
  * TO-DO: Handle error responses.
@@ -72,10 +22,8 @@ export async function getAllContent(
 ): Promise<ContentResponse> {
   const url = generateContentUrl(options);
 
-  const fetch = await getFetch();
-  const content: ContentResponse = await fetch(url.href).then((res) =>
-    res.json()
-  );
+  const res = await fetch(url.href);
+  const content = await (res.json() as Promise<ContentResponse>);
 
   const canTrack = options.canTrack !== false;
   if (