+++ packages/sdks/src/functions/get-fetch.ts
@@ -1,14 +1,18 @@
 import { getGlobalThis } from './get-global-this.js';
 
-export async function getFetch(): Promise<typeof global.fetch> {
+function getFetch(): typeof global.fetch {
   const globalFetch: typeof global.fetch = getGlobalThis().fetch;
 
-  if (typeof globalFetch === 'undefined' && typeof global !== 'undefined') {
-    const nodeFetch = import('node-fetch').then((d) => d.default) as Promise<
-      typeof global.fetch
-    >;
-    return (nodeFetch as any).default || nodeFetch;
+  if (typeof globalFetch === 'undefined') {
+    console.warn(
+      `Builder SDK could not find a global fetch function. Make sure you have a polyfill for fetch in your project. 
+      For more information, read https://github.com/BuilderIO/this-package-uses-fetch`
+    );
+
+    throw new Error('Builder SDK could not find a global `fetch` function');
   }
 
-  return (globalFetch as any).default || globalFetch;
+  return globalFetch;
 }
+
+export const fetch = getFetch();