+++ packages/sdks/src/helpers/css.ts
@@ -1,16 +1,23 @@
 import { camelToKebabCase } from '../functions/camel-to-kebab-case.js';
+import { checkIsDefined } from './nullable.js';
 
-export const convertStyleMapToCSS = (
+export const convertStyleMapToCSSArray = (
   style: Partial<CSSStyleDeclaration>
-): string => {
+): string[] => {
   const cssProps = Object.entries(style).map(([key, value]) => {
     if (typeof value === 'string') {
       return `${camelToKebabCase(key)}: ${value};`;
+    } else {
+      return undefined;
     }
   });
 
-  return cssProps.join('\n');
+  return cssProps.filter(checkIsDefined);
 };
+
+export const convertStyleMapToCSS = (
+  style: Partial<CSSStyleDeclaration>
+): string => convertStyleMapToCSSArray(style).join('\n');
 
 export const createCssClass = ({
   mediaQuery,